Stabilized circular dependencies.
