+++ src/parser.ts
@@ -1,0 +1,7 @@
+import { Component } from "./components.js";
+
+export class Parser extends Component {
+    constructor() {
+        super();
+    }
+}
