+++ src/parsers/restructuredtext.ts
@@ -1,0 +1,13 @@
+import { Parser } from "../parser.js";
+
+export class RestructuredtextParser extends Parser {
+    constructor() {
+        super();
+        // Constructor logic here
+    }
+
+    parse(input: string): any {
+        // Parsing logic here
+        return {}; // Placeholder for parsed output
+    }
+}
