--- src/readers/doctree.ts
@@ -1,5 +1,0 @@
-import { readers } from "../index.js";
-
-export class Reader extends readers.ReReader {
-
-}
