--- src/readers/pep.ts
@@ -1,5 +1,0 @@
-import * as standalone from "./standalone.js";
-
-export class Reader extends standalone.Reader {
-
-}
