+++ src/types.ts
@@ -1,0 +1,1 @@
+// TODO : Fill up with types that mimic classes to prevent circular dependencies (as needed)
