+++ src/writers/pojo.ts
@@ -1,0 +1,10 @@
+import { Writer } from "../writer.js";
+
+/**
+ * Class for writing HTML
+ *  
+ */
+export class POJOWriter extends Writer {
+
+
+}
